requirements.txt
Django==4.1.2
django_rest_passwordreset==1.3.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.5
